Improve robustness of assertion in test_newdep
The test playbook configuration is installing the latest OMERO 5.2 server. As
test_newdep.py is testing an upgrade using the latest version, it should not
be necessary to upgrade this test for each minor release. Instead, checking
that the version of the new server is 5.3.0 or above should be sufficient.