Fix failing import test following Java release
The Java security changes disabling anonymous cyphers have now been propagated
to the openjdk and package managers and causes CLI imports via the CLI to fail.

With OMERO 5.4.10, the workaround is to skip the upgrade check at import time.
This means the role import tests can no longer be run on older servers. It is
thus moved to a separate test file and restricted to the test hosts upgraded
to the latest version.