Modify default variables to install self-signed certificates by default
Recent changes to OMERO.py removing the ADH ciphers combined with
increased security requirements on modern platforms means that an
OMERO.server without self-signed certificates is unlikely to be able
to function correctly.
This changes proposes to set the default value of the internal variable
to true so that certificates are generated automatically on startup.